requirements.txt
tensorflow
numpy
scipy
matplotlib
opencv-python
datetime
tensorboard
tensorflow_addons
gitpython
